# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Inferna Next
  version: 0.1.0
paths:
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Me
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/users:
    get:
      tags:
      - users
      summary: List Users
      operationId: list_users_api_v1_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserOut'
                type: array
                title: Response List Users Api V1 Users Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - users
      summary: Create User
      operationId: create_user_api_v1_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/users/{user_id}:
    delete:
      tags:
      - users
      summary: Delete User
      operationId: delete_user_api_v1_users__user_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/{user_id}/password:
    put:
      tags:
      - users
      summary: Change Password
      operationId: change_password_api_v1_users__user_id__password_put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/clusters:
    get:
      tags:
      - clusters
      summary: List Clusters
      operationId: list_clusters_api_v1_clusters_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ClusterOut'
                type: array
                title: Response List Clusters Api V1 Clusters Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - clusters
      summary: Create Cluster
      operationId: create_cluster_api_v1_clusters_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/clusters/{cluster_id}:
    delete:
      tags:
      - clusters
      summary: Delete Cluster
      operationId: delete_cluster_api_v1_clusters__cluster_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: cluster_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Cluster Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/models:
    get:
      tags:
      - models
      summary: List Models
      operationId: list_models_api_v1_models_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModelOut'
                type: array
                title: Response List Models Api V1 Models Get
      security:
      - OAuth2PasswordBearer: []
  /api/v1/models/{model_id}:
    get:
      tags:
      - models
      summary: Get Model
      operationId: get_model_api_v1_models__model_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/model-instances:
    get:
      tags:
      - instances
      summary: List Instances
      operationId: list_instances_api_v1_model_instances_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InstanceOut'
                type: array
                title: Response List Instances Api V1 Model Instances Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - instances
      summary: Deploy
      operationId: deploy_api_v1_model_instances_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/model-instances/{instance_id}/stop:
    post:
      tags:
      - instances
      summary: Stop Instance
      operationId: stop_instance_api_v1_model_instances__instance_id__stop_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: instance_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Instance Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/model-instances/{instance_id}:
    delete:
      tags:
      - instances
      summary: Delete Instance
      operationId: delete_instance_api_v1_model_instances__instance_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: instance_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Instance Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ClusterCreate:
      properties:
        name:
          type: string
          maxLength: 64
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - name
      title: ClusterCreate
    ClusterOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - description
      - created_at
      title: ClusterOut
...
    DeployRequest:
      properties:
        model_id:
          type: string
          format: uuid
          title: Model Id
        cluster_id:
          type: string
          format: uuid
          title: Cluster Id
        engine:
          type: string
          enum:
          - vllm
          - sglang
          title: Engine
        profile:
          type: string
          enum:
          - latency
          - throughput
          title: Profile
        gpu_selection:
          anyOf:
          - $ref: '#/components/schemas/ManualGpuSelection'
          - type: string
            const: auto
          title: Gpu Selection
          default: auto
      type: object
      required:
      - model_id
      - cluster_id
      - engine
      - profile
      title: DeployRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InstanceOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        model_id:
          type: string
          format: uuid
          title: Model Id
        cluster_id:
          type: string
          format: uuid
          title: Cluster Id
        worker_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Worker Id
        engine:
          type: string
          title: Engine
        profile:
          type: string
          title: Profile
        gpu_indexes:
          items:
            type: integer
          type: array
          title: Gpu Indexes
        state:
          type: string
          title: State
        port:
          anyOf:
          - type: integer
          - type: 'null'
          title: Port
        error_detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Detail
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        model:
          anyOf:
          - $ref: '#/components/schemas/ModelOut'
          - type: 'null'
        worker_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Worker Name
      type: object
      required:
      - id
      - model_id
      - cluster_id
      - worker_id
      - engine
      - profile
      - gpu_indexes
      - state
      - port
      - error_detail
      - created_at
      - updated_at
      title: InstanceOut
    LoginRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - password
      title: LoginRequest
    ManualGpuSelection:
      properties:
        worker_id:
          type: string
          format: uuid
          title: Worker Id
        gpu_indexes:
          items:
            type: integer
          type: array
          minItems: 1
          title: Gpu Indexes
      type: object
      required:
      - worker_id
      - gpu_indexes
      title: ManualGpuSelection
    ModelOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        display_name:
          type: string
          title: Display Name
        category:
          type: string
          title: Category
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        params_b:
          anyOf:
          - type: number
          - type: 'null'
          title: Params B
        vram_required_mb:
          type: integer
          title: Vram Required Mb
        requires_hf_token:
          type: boolean
          title: Requires Hf Token
        license:
          anyOf:
          - type: string
          - type: 'null'
          title: License
        is_builtin:
          type: boolean
          title: Is Builtin
      type: object
      required:
      - id
      - name
      - display_name
      - category
      - description
      - params_b
      - vram_required_mb
      - requires_hf_token
      - license
      - is_builtin
      title: ModelOut
    PasswordChange:
      properties:
        password:
          type: string
          maxLength: 128
          minLength: 6
          title: Password
      type: object
      required:
      - password
      title: PasswordChange
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      title: TokenResponse
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 128
          minLength: 6
          title: Password
        role:
          type: string
          title: Role
          default: user
      type: object
      required:
      - username
      - password
      title: UserCreate
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        username:
          type: string
          title: Username
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - id
      - username
      - email
      - role
      - is_active
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login